requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.4.2
pydantic-settings>=2.0.3
google-auth-oauthlib>=1.1.0
google-auth>=2.23.3
google-auth-httplib2>=0.1.0
google-api-python-client>=2.108.0
openai>=1.3.0
pytest>=7.4.3
httpx>=0.25.1
python-dotenv>=1.0.0
email-validator>=2.1.0.post1
PyJWT>=2.0.0